optimize: optimize the logic of identifying IDE
