+++ .gitignore
@@ -6,3 +6,4 @@ cache
 .temp
 dist.zip
 .DS_Store
+.idea
