+++ packages/core/src/server/launch-editor/editor-info/linux.ts
@@ -1,0 +1,44 @@
+import { EDITOR_PROCESS_MAP } from "../type";
+
+export const COMMON_EDITORS_LINUX = {
+  cursor: 'cursor',
+  code: 'code',
+  vscodium: 'vscodium',
+  codium: 'codium',
+  webstorm: 'webstorm',
+  'webstorm.sh': 'webstorm',
+  hbuilderx: 'hbuilderx',
+  'hbuilderx.sh': 'hbuilderx',
+  atom: 'atom',
+  Brackets: 'brackets',
+  'code-insiders': 'code-insiders',
+  emacs: 'emacs',
+  gvim: 'gvim',
+  idea: 'idea',
+  'idea.sh': 'idea',
+  phpstorm: 'phpstorm',
+  'phpstorm.sh': 'phpstorm',
+  pycharm: 'pycharm',
+  'pycharm.sh': 'pycharm',
+  rubymine: 'rubymine',
+  'rubymine.sh': 'rubymine',
+  sublime_text: 'subl',
+  vim: 'vim',
+  goland: 'goland',
+  'goland.sh': 'goland',
+  rider: 'rider',
+  'rider.sh': 'rider',
+};
+
+export const EDITOR_PROCESS_MAP_LINUX: EDITOR_PROCESS_MAP = {
+  code: ['code'],
+  code_insiders: ['code-insiders'],
+  webstorm: ['webstorm.sh'],
+  cursor: ['cursor'],
+  atom: ['atom'],
+  hbuilder: ['hbuilderx'],
+  phpstorm: ['phpstorm.sh'],
+  pycharm: ['pycharm.sh'],
+  idea: ['idea.sh'],
+  codium: ['vscodium'],
+}
