+++ packages/core/src/server/launch-editor/editor-info/windows.ts
@@ -1,0 +1,45 @@
+import { EDITOR_PROCESS_MAP } from "../type";
+
+export const COMMON_EDITORS_WIN: { [key: string]: string } = {
+  'Cursor.exe': '',
+  'Code.exe': '',
+  'Code - Insiders.exe': '',
+  'VSCodium.exe': '',
+  'webstorm.exe': '',
+  'webstorm64.exe': '',
+  'HBuilderX.exe': '',
+  'HBuilder.exe': '',
+  'HBuilderX64.exe': '',
+  'HBuilder64.exe': '',
+  'Brackets.exe': '',
+  'atom.exe': '',
+  'sublime_text.exe': '',
+  'notepad++.exe': '',
+  'clion.exe': '',
+  'clion64.exe': '',
+  'idea.exe': '',
+  'idea64.exe': '',
+  'phpstorm.exe': '',
+  'phpstorm64.exe': '',
+  'pycharm.exe': '',
+  'pycharm64.exe': '',
+  'rubymine.exe': '',
+  'rubymine64.exe': '',
+  'goland.exe': '',
+  'goland64.exe': '',
+  'rider.exe': '',
+  'rider64.exe': '',
+};
+
+export const EDITOR_PROCESS_MAP_WIN: EDITOR_PROCESS_MAP = {
+  code: ['Code.exe'],
+  code_insiders: ['Code - Insiders.exe'],
+  webstorm: ['webstorm.exe', 'webstorm64.exe'],
+  cursor: ['Cursor.exe'],
+  atom: ['atom.exe'],
+  hbuilder: ['HBuilderX.exe', 'HBuilder.exe', 'HBuilderX64.exe', 'HBuilder64.exe'],
+  phpstorm: ['phpstorm.exe', 'phpstorm64.exe'],
+  pycharm: ['pycharm.exe', 'pycharm64.exe'],
+  idea: ['idea.exe', 'idea64.exe'],
+  codium: ['VSCodium.exe'],
+}
