+++ packages/core/src/server/launch-editor/type.ts
@@ -1,0 +1,6 @@
+import { Editor } from "../../shared";
+
+export type Platform = 'darwin' | 'linux' | 'win32';
+export type EDITOR_PROCESS_MAP = {
+  [key in Editor]: string[];
+}
