--- packages/core/types/server/launch-editor.d.ts
@@ -1,3 +1,0 @@
-import { Editor, IDEOpenMethod } from '../shared';
-declare function launchEditor(fileName: string, lineNumber: unknown, colNumber: unknown, _editor?: Editor, ideOpenMethod?: IDEOpenMethod, pathFormat?: string | string[]): void;
-export default launchEditor;
