+++ packages/core/types/server/launch-editor/editor-info/index.d.ts
@@ -1,0 +1,3 @@
+import { EDITOR_PROCESS_MAP, Platform } from '../type';
+export declare const COMMON_EDITORS_MAP: Record<Platform, Record<string, string>>;
+export declare const COMMON_EDITOR_PROCESS_MAP: Record<Platform, EDITOR_PROCESS_MAP>;
