+++ packages/core/types/server/launch-editor/editor-info/linux.d.ts
@@ -1,0 +1,31 @@
+import { EDITOR_PROCESS_MAP } from "../type";
+export declare const COMMON_EDITORS_LINUX: {
+    cursor: string;
+    code: string;
+    vscodium: string;
+    codium: string;
+    webstorm: string;
+    'webstorm.sh': string;
+    hbuilderx: string;
+    'hbuilderx.sh': string;
+    atom: string;
+    Brackets: string;
+    'code-insiders': string;
+    emacs: string;
+    gvim: string;
+    idea: string;
+    'idea.sh': string;
+    phpstorm: string;
+    'phpstorm.sh': string;
+    pycharm: string;
+    'pycharm.sh': string;
+    rubymine: string;
+    'rubymine.sh': string;
+    sublime_text: string;
+    vim: string;
+    goland: string;
+    'goland.sh': string;
+    rider: string;
+    'rider.sh': string;
+};
+export declare const EDITOR_PROCESS_MAP_LINUX: EDITOR_PROCESS_MAP;
