+++ packages/core/types/server/launch-editor/editor-info/mac.d.ts
@@ -1,0 +1,30 @@
+import { EDITOR_PROCESS_MAP } from '../type';
+export declare const COMMON_EDITORS_OSX: {
+    '/Cursor.app/Contents/MacOS/Cursor': string;
+    '/Visual Studio Code.app/Contents/MacOS/Electron': string;
+    '/Visual Studio Code - Insiders.app/Contents/MacOS/Electron': string;
+    '/VSCodium.app/Contents/MacOS/Electron': string;
+    '/WebStorm.app/Contents/MacOS/webstorm': string;
+    '/HBuilderX.app/Contents/MacOS/HBuilderX': string;
+    '/Atom.app/Contents/MacOS/Atom': string;
+    '/Atom Beta.app/Contents/MacOS/Atom Beta': string;
+    '/Brackets.app/Contents/MacOS/Brackets': string;
+    '/Sublime Text.app/Contents/MacOS/Sublime Text': string;
+    '/Sublime Text.app/Contents/MacOS/sublime_text': string;
+    '/Sublime Text 2.app/Contents/MacOS/Sublime Text 2': string;
+    '/Sublime Text Dev.app/Contents/MacOS/Sublime Text': string;
+    '/PhpStorm.app/Contents/MacOS/phpstorm': string;
+    '/PyCharm.app/Contents/MacOS/pycharm': string;
+    '/PyCharm CE.app/Contents/MacOS/pycharm': string;
+    '/IntelliJ IDEA.app/Contents/MacOS/idea': string;
+    '/IntelliJ IDEA Ultimate.app/Contents/MacOS/idea': string;
+    '/IntelliJ IDEA Community Edition.app/Contents/MacOS/idea': string;
+    '/Zed.app/Contents/MacOS/zed': string;
+    '/GoLand.app/Contents/MacOS/goland': string;
+    '/AppCode.app/Contents/MacOS/appcode': string;
+    '/CLion.app/Contents/MacOS/clion': string;
+    '/RubyMine.app/Contents/MacOS/rubymine': string;
+    '/MacVim.app/Contents/MacOS/MacVim': string;
+    '/Rider.app/Contents/MacOS/rider': string;
+};
+export declare const EDITOR_PROCESS_MAP_OSX: EDITOR_PROCESS_MAP;
