+++ packages/core/types/server/launch-editor/editor-info/windows.d.ts
@@ -1,0 +1,5 @@
+import { EDITOR_PROCESS_MAP } from "../type";
+export declare const COMMON_EDITORS_WIN: {
+    [key: string]: string;
+};
+export declare const EDITOR_PROCESS_MAP_WIN: EDITOR_PROCESS_MAP;
