+++ packages/core/types/server/launch-editor/get-args.d.ts
@@ -1,0 +1,9 @@
+export declare function getArguments(params: {
+    processName: string;
+    fileName: string;
+    lineNumber: string | number;
+    colNumber: string | number;
+    workspace: string | null;
+    openWindowParams: string;
+    pathFormat?: string | string[];
+}): string[];
