+++ packages/core/types/server/launch-editor/guess.d.ts
@@ -1,0 +1,2 @@
+import { Editor } from '../../shared';
+export declare function guessEditor(_editor?: Editor): string[] | null[];
