+++ packages/core/types/server/launch-editor/index.d.ts
@@ -1,0 +1,3 @@
+import { Editor, IDEOpenMethod } from '../../shared';
+declare function launchEditor(fileName: string, lineNumber: number, colNumber: number, _editor?: Editor, ideOpenMethod?: IDEOpenMethod, pathFormat?: string | string[]): void;
+export default launchEditor;
