+++ packages/core/types/server/launch-editor/type.d.ts
@@ -1,0 +1,5 @@
+import { Editor } from "../../shared";
+export type Platform = 'darwin' | 'linux' | 'win32';
+export type EDITOR_PROCESS_MAP = {
+    [key in Editor]: string[];
+};
